DEMRUM-4776: address PR review feedback
* Fix build: add @_spi(SplunkTesting) import for waitUntil
* Add DocC comment to isNavigationControllerManaged
* Add MARK sections in NavigationModel for pending/managed targets
* Remove duplicated screen-update logic in processNavigationControllerWillShow
* Fix non-atomic double actor read in completeNavigationControllerTransition
* Restructure test helpers: split mocks into Mocks/, move fixture
  helpers to NavigationStreamFixture extension methods
* Rename test file/class to NavigationControllerSwizzlingTests
* Add CHANGELOG entry for navigation controller transitions
