+++ SplunkNavigation/Sources/SplunkNavigation/Navigation.swift
@@ -277,6 +277,14 @@ public final class Navigation: Sendable {
         return moduleEnabled && trackingEnabled
     }
 
+    /// Checks whether the event belongs to a view controller whose lifecycle is managed
+    /// by a `UINavigationController` transition.
+    ///
+    /// When a navigation controller push or pop is in flight, the child controller also fires
+    /// its own `viewDidLoad` / `viewDidAppear` events. Processing those independently would
+    /// create duplicate screen-name updates and navigation spans. This guard suppresses the
+    /// redundant lifecycle events so that only the navigation-controller transition path
+    /// (`willShow` / `didShow`) drives the screen change.
     private func isNavigationControllerManaged(event: NavigationActionEvent) async -> Bool {
         switch event.type {
         case .viewDidAppear,
