+++ SplunkNavigation/Tests/SplunkNavigationTests/Testing Support/Mocks/MockNavigationActionEvent.swift
@@ -1,0 +1,30 @@
+//
+/*
+Copyright 2026 Splunk Inc.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+internal import CiscoSwizzling
+import Foundation
+
+@testable import SplunkNavigation
+
+struct MockNavigationActionEvent: NavigationActionEvent {
+    let timestamp: Date
+    let type: NavigationActionEventType
+    let navigationControllerIdentifier: ObjectIdentifier?
+    let controllerTypeName: String
+    let controllerIdentifier: ObjectIdentifier
+    var viewFrame: CGRect?
+}
