+++ SplunkNavigation/Tests/SplunkNavigationTests/Testing Support/Mocks/MockNavigationEventStreamProvider.swift
